refactor: 문자열 파싱 함수 분리하기
